requirements.txt
Flask==1.0.2
flask-mongoengine==0.9.5
docopt==0.6.2
colorama==0.4.1
psutil==5.6.3
pyyaml==5.1.2
gunicorn==19.9.0
gevent==1.4.0
passlib==1.7.1
python-jose==3.0.1
flask-restful==0.3.7